moved FS mutation tests to separate file
